Starting a new project: "Forum"
Forum will be easy to use, secure, and fast. I'm also planning on adding
scriptable components that will use admin-configurable JavaScript when
certain actions are being performed ("run this to see if the user can
actually post").

This initial commit contains a stub of a README, enough code to connect
to a Couchbase server and securely display a 404 page, profiling support
(go build -tags profile), and a 1024-bit TLS key valid until the end of
the year that can be used for local testing (it's valid on 127.0.0.1 and
localhost. You can generate your own using the script in Go's crypto/tls
package).